refactor: split repository responsibilities and update dependencies
Split the monolithic RequestItemsRepository into separate ItemsRepository (for product retrieval) and OrdersRepository (for order submission) to improve separation of concerns. Update the dependency injection bindings and refactor the RequestItemsViewModel to depend on the new repository structure, enhancing code maintainability and architectural clarity.
