+++ app/app/src/main/java/com/lojasocial/app/di/RepositoryModule.kt
@@ -1,7 +1,9 @@
 package com.lojasocial.app.di
 
-import com.lojasocial.app.repository.RequestItemsRepository
-import com.lojasocial.app.repository.RequestItemsRepositoryImpl
+import com.lojasocial.app.repository.ItemsRepository
+import com.lojasocial.app.repository.ItemsRepositoryImpl
+import com.lojasocial.app.repository.OrdersRepository
+import com.lojasocial.app.repository.OrdersRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -14,7 +16,13 @@ abstract class RepositoryModule {
 
     @Binds
     @Singleton
-    abstract fun bindRequestItemsRepository(
-        requestItemsRepositoryImpl: RequestItemsRepositoryImpl
-    ): RequestItemsRepository
+    abstract fun bindItemsRepository(
+        itemsRepositoryImpl: ItemsRepositoryImpl
+    ): ItemsRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindOrdersRepository(
+        ordersRepositoryImpl: OrdersRepositoryImpl
+    ): OrdersRepository
 }
